fix bug comparing uint and ulong

+++ libraries/RCReceiver/RCReceiver.h
@@ -28,7 +28,7 @@ class RC {
     float m_offsetmax[CHANNEL_NB];
     float m_offsetmin[CHANNEL_NB];
 
-    unsigned int m_last_interrupt;
+    unsigned long m_last_interrupt;
